# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Cost-RAG 工程造价咨询智能RAG系统 API
  description: |
    Cost-RAG是基于成本优化的检索增强生成系统，专为工程造价咨询行业设计。
    提供文档处理、智能问答、成本估算和多项目对比等核心功能。

    ## 主要特性
    - 🔍 支持多格式文档处理和智能检索
    - 💰 14级分部分项层级成本计算
    - 📊 多项目对比分析和相似性评估
    - 🤖 基于RAG的智能问答系统
    - 🚀 高性能异步处理架构

    ## 认证方式
    使用JWT Bearer Token进行认证，在请求头中添加：
    ```
    Authorization: Bearer <your_jwt_token>
    ```

    ## 错误处理
    所有错误响应都遵循统一格式：
    ```json
    {
      "success": false,
      "error": {
        "code": "ERROR_CODE",
        "message": "错误描述",
        "details": []
      }
    }
    ```
  version: 1.0.0
  contact:
    name: Cost-RAG开发团队
    email: support@cost-rag.com
    url: https://cost-rag.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000/api/v1
    description: 开发环境
  - url: https://api.cost-rag.com/api/v1
    description: 生产环境
paths:
...
  /estimates:
    post:
      tags:
        - 成本估算
      summary: 创建成本估算
      description: |
        基于项目参数创建新的成本估算，使用14级层级递归计算算法。
        系统将自动选择合适的成本模板并进行质量调整。
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEstimateRequest'
      responses:
        '201':
          description: 估算创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateEstimateResponse'
        '422':
          $ref: '#/components/responses/ValidationError'

    get:
      tags:
        - 成本估算
      summary: 获取估算列表
      description: 分页获取成本估算列表，支持项目类型和状态过滤
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - name: project_type
          in: query
          description: 项目类型过滤
          schema:
            type: string
            enum: [office, residential, commercial, mixed]
        - name: status
          in: query
          description: 估算状态过滤
          schema:
            type: string
            enum: [draft, processing, completed, failed]
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EstimateListResponse'
  /estimates/{estimate_id}:
    get:
      tags:
        - 成本估算
      summary: 获取估算详情
      description: 获取成本估算的详细信息，包括完整的14级分部分项分解
      security:
        - bearerAuth: []
      parameters:
        - name: estimate_id
          in: path
          required: true
          description: 估算ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EstimateDetailResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'

    put:
      tags:
        - 成本估算
      summary: 更新估算
      description: 更新成本估算的项目参数，系统将重新计算成本分解
      security:
        - bearerAuth: []
      parameters:
        - name: estimate_id
          in: path
          required: true
          description: 估算ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEstimateRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EstimateDetailResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '422':
          $ref: '#/components/responses/ValidationError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CreateEstimateRequest:
      type: object
      required:
        - project_name
        - project_type
        - area
      properties:
        project_name:
          type: string
          description: 项目名称
          example: "商业综合体项目"
          minLength: 2
          maxLength: 255
        project_type:
          type: string
          enum: [office, residential, commercial, mixed]
          description: 项目类型
        area:
          type: number
          format: float
          description: 建筑面积(平方米)
          minimum: 1
          example: 50000.0
        floors:
          type: integer
          description: 楼层数
          minimum: 1
          maximum: 200
          example: 25
        quality_level:
          type: string
          enum: [low, medium, high]
          default: medium
          description: 质量等级
        location:
          type: string
          description: 项目位置
          example: "北京市朝阳区"
          maxLength: 255
        template_id:
          type: string
          format: uuid
          description: 指定的成本模板ID(可选)
        description:
          type: string
          description: 项目描述
          maxLength: 1000
    CreateEstimateResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            estimate_id:
              type: string
              format: uuid
            project_name:
              type: string
            status:
              type: string
              enum: [draft, processing, completed, failed]
            estimated_completion_time:
              type: integer
              description: 预估完成时间(秒)
            created_at:
              type: string
              format: date-time
    EstimateListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            estimates:
              type: array
              items:
                $ref: '#/components/schemas/EstimateSummary'
            pagination:
              $ref: '#/components/schemas/PaginationInfo'
    EstimateSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        project_name:
          type: string
        project_type:
          type: string
        area:
          type: number
          format: float
        status:
          type: string
        total_cost:
          type: number
          format: float
          description: 总造价(元)
        unit_cost:
          type: number
          format: float
          description: 单方造价(元/平米)
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    EstimateDetailResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            estimate:
              $ref: '#/components/schemas/EstimateSummary'
            breakdown:
              $ref: '#/components/schemas/CostBreakdown'
            template_used:
              type: object
              properties:
                template_id:
                  type: string
                  format: uuid
                template_name:
                  type: string
                region:
                  type: string
            validation_results:
              $ref: '#/components/schemas/ValidationResults'
            calculation_metadata:
              type: object
              properties:
                calculation_method:
                  type: string
                  example: "hierarchical_14_level"
                processing_time:
                  type: number
                  format: float
                quality_adjustments:
                  type: object
                  additionalProperties:
                    type: number
    CostBreakdown:
      type: object
      properties:
        total_cost:
          type: number
          format: float
          description: 总造价(元)
        unit_cost:
          type: number
          format: float
          description: 单方造价(元/平米)
        primary_sections:
          type: array
          items:
            $ref: '#/components/schemas/PrimarySection'
        secondary_sections:
          type: array
          items:
            $ref: '#/components/schemas/SecondarySection'
    PrimarySection:
      type: object
      properties:
        section_code:
          type: string
          example: "1.0"
        section_name:
          type: string
          example: "土石方工程"
        unit_price:
          type: number
          format: float
          description: 单方造价(元/平米)
        total_cost:
          type: number
          format: float
          description: 合价(元)
        cost_ratio:
          type: number
          format: float
          description: 占总造价比例
        secondary_sections_count:
          type: integer
          description: 包含的二级分部数量
    SecondarySection:
      type: object
      properties:
        section_code:
          type: string
          example: "2.1"
        section_name:
          type: string
          example: "土方开挖"
        primary_section_code:
          type: string
          example: "2.0"
        unit_price:
          type: number
          format: float
        total_cost:
          type: number
          format: float
        cost_breakdown:
          type: object
          properties:
            material_cost:
              type: number
              format: float
            labor_cost:
              type: number
              format: float
            equipment_cost:
              type: number
              format: float
    UpdateEstimateRequest:
      type: object
      properties:
        project_name:
          type: string
          minLength: 2
          maxLength: 255
        area:
          type: number
          format: float
          minimum: 1
        quality_level:
          type: string
          enum: [low, medium, high]
        floors:
          type: integer
          minimum: 1
          maximum: 200
        location:
          type: string
          maxLength: 255
        description:
          type: string
          maxLength: 1000

    # 多项目对比相关
...
    PaginationInfo:
      type: object
      properties:
        page:
          type: integer
          minimum: 1
          example: 1
        size:
          type: integer
          minimum: 1
          maximum: 100
          example: 20
        total:
          type: integer
          minimum: 0
          example: 150
        pages:
          type: integer
          minimum: 0
          example: 8
        has_next:
          type: boolean
        has_prev:
          type: boolean
    ValidationResults:
      type: object
      properties:
        is_valid:
          type: boolean
        errors:
          type: array
          items:
            type: string
        warnings:
          type: array
          items:
            type: string
        mathematical_validation:
          type: object
          properties:
            section_14_validation:
              type: boolean
            hierarchy_validations:
              type: array
              items:
                type: object
                properties:
                  section_code:
                    type: string
                  is_valid:
                    type: boolean
                  expected_value:
                    type: number
                  actual_value:
                    type: number
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: "VALIDATION_ERROR"
            message:
              type: string
              example: "请求参数验证失败"
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
        timestamp:
          type: string
          format: date-time
        request_id:
          type: string
          format: uuid
  parameters:
    PageParam:
      name: page
      in: query
      description: 页码，从1开始
      schema:
        type: integer
        minimum: 1
        default: 1
    SizeParam:
      name: size
      in: query
      description: 每页数量，最大100
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  responses:
...
    NotFoundError:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ValidationError:
      description: 数据验证失败
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...